Create separate file for enum to string converters. Move all exceptions out of include folder.
