+++ c_abi/src/c_message.cpp
@@ -1,6 +1,7 @@
 #include "c_message.hpp"
 #include "c_thread_channel.hpp"
 #include "c_errors.hpp"
+#include "application/enum_converters.hpp"
 #include "nlohmann/json.hpp"
 #include "message.hpp"
 #include "thread_message.hpp"
