+++ src/application/message_service.cpp
@@ -1,5 +1,6 @@
 #include "message_service.hpp"
 #include "chat_service.hpp"
+#include "enum_converters.hpp"
 #include "domain/message_entity.hpp"
 #include "domain/quotes.hpp"
 #include "infra/pubnub.hpp"
